requirements/local.txt
# Archivo requirements local

## includes the base.txt requirements file
-r base.txt

django-extensions==1.7.7
ipython==5.3.0
Werkzeug==0.11.15

# Sphinx
Sphinx==1.5.2
sphinx-autobuild==0.6.0

# Linters
flake8==3.3.0
autopep8==1.2.4

# Tests
coverage==4.3.4
isort==4.2.5
tox==2.6.0
